requirements.txt
alembic==1.17.2
blinker==1.9.0
build==1.3.0
click==8.3.1
coverage==7.12.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pip-tools==7.5.2
pluggy==1.6.0
psycopg2==2.9.11
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==9.0.1
python-decouple==3.8
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
